Move the code for the view Python module to the dedicated directory
Each Python extension module should use a dedicated directory.  It is also the case for the embedded extension module _modmesh_view.
