+++ cpp/modmesh/CMakeLists.txt
@@ -4,6 +4,7 @@
 cmake_minimum_required(VERSION 3.16)
 
 add_subdirectory(python/wrapper/modmesh)
+add_subdirectory(python/wrapper/view)
 add_subdirectory(view)
 
 set(MODMESH_HEADERS
