+++ cpp/modmesh/python/wrapper/modmesh/wrap_ConcreteBuffer.cpp
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/python/python.hpp> // Must be the first include.
+#include <modmesh/python/wrapper/modmesh/modmesh.hpp> // Must be the first include.
 #include <modmesh/modmesh.hpp>
 
 namespace modmesh
