+++ cpp/modmesh/view/PythonInterpreter.cpp
@@ -121,124 +121,6 @@ _set_modmesh_path())"""";
     py::exec("modmesh.view = _modmesh_view");
 }
 
-namespace python
-{
-
-// clang-format off
-class
-MODMESH_PYTHON_WRAPPER_VISIBILITY
-WrapR3DWidget
-    // clang-format on
-    : public WrapBase<WrapR3DWidget, R3DWidget>
-{
-
-    friend root_base_type;
-
-    WrapR3DWidget(pybind11::module & mod, char const * pyname, char const * pydoc)
-        : root_base_type(mod, pyname, pydoc)
-    {
-
-        namespace py = pybind11;
-
-#define DECL_QVECTOR3D_PROPERTY(NAME, GETTER, SETTER)          \
-    .def_property(                                             \
-        #NAME,                                                 \
-        [](wrapped_type & self)                                \
-        {                                                      \
-            QVector3D const v = self.camera()->GETTER();       \
-            return py::make_tuple(v.x(), v.y(), v.z());        \
-        },                                                     \
-        [](wrapped_type & self, std::vector<double> const & v) \
-        {                                                      \
-            double const x = v.at(0);                          \
-            double const y = v.at(1);                          \
-            double const z = v.at(2);                          \
-            self.camera()->SETTER(QVector3D(x, y, z));         \
-        })
-
-        (*this)
-            DECL_QVECTOR3D_PROPERTY(position, position, setPosition)
-                DECL_QVECTOR3D_PROPERTY(up_vector, upVector, setUpVector)
-                    DECL_QVECTOR3D_PROPERTY(view_center, viewCenter, setViewCenter)
-            //
-            ;
-
-#undef DECL_QVECTOR3D_PROPERTY
-    }
-
-}; /* end class WrapR3DWidget */
-
-class
-    MODMESH_PYTHON_WRAPPER_VISIBILITY
-        WrapRApplication
-    // clang-format on
-    : public WrapBase<WrapRApplication, RApplication>
-{
-
-    friend root_base_type;
-
-    WrapRApplication(pybind11::module & mod, char const * pyname, char const * pydoc)
-        : root_base_type(mod, pyname, pydoc)
-    {
-
-        namespace py = pybind11;
-
-        (*this)
-            .def_property_readonly_static(
-                "instance",
-                [](py::object const &)
-                {
-                    return RApplication::instance();
-                })
-            .def_property_readonly(
-                "viewer",
-                [](wrapped_type & self)
-                {
-                    return self.main()->viewer();
-                })
-            //
-            ;
-    }
-
-}; /* end class WrapRApplication */
-
-namespace detail
-{
-
-static void update_appmesh(std::shared_ptr<StaticMesh> const & mesh)
-{
-    RScene * scene = RApplication::instance()->main()->viewer()->scene();
-    for (Qt3DCore::QNode * child : scene->childNodes())
-    {
-        if (typeid(*child) == typeid(RStaticMesh))
-        {
-            child->deleteLater();
-        }
-    }
-    new RStaticMesh(mesh, scene);
-}
-
-} /* end namespace detail */
-
-} /* end namespace python */
-
 } /* end namespace modmesh */
-
-PYBIND11_EMBEDDED_MODULE(_modmesh_view, mod)
-{
-    using namespace modmesh;
-    using namespace modmesh::python;
-    namespace py = pybind11;
-
-    mod
-        .def("show", &modmesh::python::detail::update_appmesh, py::arg("mesh"))
-        //
-        ;
-
-    WrapR3DWidget::commit(mod, "R3DWidget", "R3DWidget");
-    WrapRApplication::commit(mod, "RApplication", "RApplication");
-
-    mod.attr("app") = py::cast(RApplication::instance());
-}
 
 // vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
