+++ cpp/modmesh/view/PythonInterpreter.hpp
@@ -35,6 +35,7 @@
 namespace modmesh
 {
 
+// FIXME: Move to python namespace and the python directory.
 class PythonInterpreter
 {
